requirements.txt
pycrypto==2.6.1
envyaml==1.1.201202
ffmpeg-python==0.2.0
